Added support for min/max for decimal
